+++ Ui/Areas/Admin/Controllers/AdminHomeController.cs
@@ -1,4 +1,5 @@
 ﻿using Entities.Dto;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 using System.Text;
@@ -7,6 +8,7 @@ using Ui.Helper;
 namespace Ui.Areas.Admin.Controllers
 {
     [ApiExplorerSettings(IgnoreApi = true)]
+    [Authorize(Roles = "Admin")]
     [Area("Admin")]
     [Route("Admin/AdminHome")]
     public class AdminHomeController : Controller
@@ -21,14 +23,18 @@ namespace Ui.Areas.Admin.Controllers
         [Route("Index")]
         public async Task<IActionResult> Index()
         {
-            var client = _httpClientFactory.CreateClient();
-            var responseMessage = await client.GetAsync("https://localhost:7140/api/Home/GetAll");
-            if (responseMessage.IsSuccessStatusCode)
+            var token = User.Claims.FirstOrDefault(x => x.Type == "accessToken")?.Value;
+            if (token != null)
             {
-                var jsonData = await responseMessage.Content.ReadAsStringAsync();
-                var apiResponse = JsonConvert.DeserializeObject<ApiResponse<List<HomeDto>>>(jsonData);
-                var values = apiResponse?.Data;
-                return View(values);
+                var client = _httpClientFactory.CreateClient();
+                var responseMessage = await client.GetAsync("https://localhost:7140/api/Home/GetAll");
+                if (responseMessage.IsSuccessStatusCode)
+                {
+                    var jsonData = await responseMessage.Content.ReadAsStringAsync();
+                    var apiResponse = JsonConvert.DeserializeObject<ApiResponse<List<HomeDto>>>(jsonData);
+                    var values = apiResponse?.Data;
+                    return View(values);
+                }
             }
             return View();
         }
@@ -44,13 +50,18 @@ namespace Ui.Areas.Admin.Controllers
         [Route("CreateHome")]
         public async Task<IActionResult> CreateHome(CreateHomeDto createHomeDto)
         {
-            var client = _httpClientFactory.CreateClient();
-            var jsonData = JsonConvert.SerializeObject(createHomeDto);
-            StringContent stringContent = new StringContent(jsonData, Encoding.UTF8, "application/json");
-            var responseMessage = await client.PostAsync("https://localhost:7140/api/Home/AddHome", stringContent);
-            if (responseMessage.IsSuccessStatusCode)
+            var token = User.Claims.FirstOrDefault(x => x.Type == "accessToken")?.Value;
+            if (token != null)
             {
-                return RedirectToAction("Index", "AdminHome", new { area = "Admin" });
+                var client = _httpClientFactory.CreateClient();
+                var jsonData = JsonConvert.SerializeObject(createHomeDto);
+                StringContent stringContent = new StringContent(jsonData, Encoding.UTF8, "application/json");
+                client.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", token);
+                var responseMessage = await client.PostAsync("https://localhost:7140/api/Home/AddHome", stringContent);
+                if (responseMessage.IsSuccessStatusCode)
+                {
+                    return RedirectToAction("Index", "AdminHome", new { area = "Admin" });
+                }
             }
             return View();
         }
@@ -58,11 +69,16 @@ namespace Ui.Areas.Admin.Controllers
         [Route("DeleteHome/{id}")]
         public async Task<IActionResult> DeleteHome(int id)
         {
-            var client = _httpClientFactory.CreateClient();
-            var responseMessage = await client.DeleteAsync($"https://localhost:7140/api/Home/DeleteHome/{id}");
-            if (responseMessage.IsSuccessStatusCode)
+            var token = User.Claims.FirstOrDefault(x => x.Type == "accessToken")?.Value;
+            if (token != null)
             {
-                return RedirectToAction("Index", "AdminHome", new { area = "Admin" });
+                var client = _httpClientFactory.CreateClient();
+                client.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", token);
+                var responseMessage = await client.DeleteAsync($"https://localhost:7140/api/Home/DeleteHome/{id}");
+                if (responseMessage.IsSuccessStatusCode)
+                {
+                    return RedirectToAction("Index", "AdminHome", new { area = "Admin" });
+                }
             }
             return RedirectToAction("Index", "AdminHome", new { area = "Admin" });
         }
@@ -71,16 +87,19 @@ namespace Ui.Areas.Admin.Controllers
         [Route("UpdateHome/{id}")]
         public async Task<IActionResult> UpdateHome(int id)
         {
-
-            var client = _httpClientFactory.CreateClient();
-            var responseMessage = await client.GetAsync($"https://localhost:7140/api/Home/GetHomeById?id={id}");
-            if (responseMessage.IsSuccessStatusCode)
+            var token = User.Claims.FirstOrDefault(x => x.Type == "accessToken")?.Value;
+            if (token != null)
             {
-                var jsonData = await responseMessage.Content.ReadAsStringAsync();
-                Console.WriteLine($"API Response: {jsonData}");
-                var apiResponse = JsonConvert.DeserializeObject<ApiResponse<HomeDto>>(jsonData);
-                var values = apiResponse?.Data;
-                return View(values);
+                var client = _httpClientFactory.CreateClient();
+                var responseMessage = await client.GetAsync($"https://localhost:7140/api/Home/GetHomeById?id={id}");
+                if (responseMessage.IsSuccessStatusCode)
+                {
+                    var jsonData = await responseMessage.Content.ReadAsStringAsync();
+                    Console.WriteLine($"API Response: {jsonData}");
+                    var apiResponse = JsonConvert.DeserializeObject<ApiResponse<HomeDto>>(jsonData);
+                    var values = apiResponse?.Data;
+                    return View(values);
+                }
             }
             return View();
         }
@@ -90,13 +109,18 @@ namespace Ui.Areas.Admin.Controllers
         [Route("UpdateHome/{id}")]
         public async Task<IActionResult> UpdateHome(HomeDto homeDto)
         {
-            var client = _httpClientFactory.CreateClient();
-            var jsonData = JsonConvert.SerializeObject(homeDto);
-            StringContent stringContent = new StringContent(jsonData, Encoding.UTF8, "application/json");
-            var responseMessage = await client.PostAsync("https://localhost:7140/api/Home/UpdateHome", stringContent);
-            if (responseMessage.IsSuccessStatusCode)
+            var token = User.Claims.FirstOrDefault(x => x.Type == "accessToken")?.Value;
+            if (token != null)
             {
-                return RedirectToAction("Index", "AdminHome", new { area = "Admin" });
+                var client = _httpClientFactory.CreateClient();
+                var jsonData = JsonConvert.SerializeObject(homeDto);
+                StringContent stringContent = new StringContent(jsonData, Encoding.UTF8, "application/json");
+                client.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", token);
+                var responseMessage = await client.PostAsync("https://localhost:7140/api/Home/UpdateHome", stringContent);
+                if (responseMessage.IsSuccessStatusCode)
+                {
+                    return RedirectToAction("Index", "AdminHome", new { area = "Admin" });
+                }
             }
             return View();
         }
