+++ Ui/Areas/Admin/Controllers/AdminLocationController.cs
@@ -52,13 +52,18 @@ namespace Ui.Areas.Admin.Controllers
         [Route("CreateLocation")]
         public async Task<IActionResult> CreateLocation(CreateLocationDto createLocationDto)
         {
-            var client = _httpClientFactory.CreateClient();
-            var jsonData = JsonConvert.SerializeObject(createLocationDto);
-            StringContent stringContent = new StringContent(jsonData, Encoding.UTF8, "application/json");
-            var responseMessage = await client.PostAsync("https://localhost:7140/api/Location/AddLocation", stringContent);
-            if (responseMessage.IsSuccessStatusCode)
+            var token = User.Claims.FirstOrDefault(x => x.Type == "accessToken")?.Value;
+            if (token != null)
             {
-                return RedirectToAction("Index", "AdminLocation", new { area = "Admin" });
+                var client = _httpClientFactory.CreateClient();
+                var jsonData = JsonConvert.SerializeObject(createLocationDto);
+                StringContent stringContent = new StringContent(jsonData, Encoding.UTF8, "application/json");
+                client.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", token);
+                var responseMessage = await client.PostAsync("https://localhost:7140/api/Location/AddLocation", stringContent);
+                if (responseMessage.IsSuccessStatusCode)
+                {
+                    return RedirectToAction("Index", "AdminLocation", new { area = "Admin" });
+                }
             }
             return View();
         }
@@ -66,11 +71,16 @@ namespace Ui.Areas.Admin.Controllers
         [Route("DeleteLocation/{id}")]
         public async Task<IActionResult> DeleteLocation(int id)
         {
-            var client = _httpClientFactory.CreateClient();
-            var responseMessage = await client.DeleteAsync($"https://localhost:7140/api/Location/DeleteLocation?id={id}");
-            if (responseMessage.IsSuccessStatusCode)
+            var token = User.Claims.FirstOrDefault(x => x.Type == "accessToken")?.Value;
+            if (token != null)
             {
-                return RedirectToAction("Index", "AdminLocation", new { area = "Admin" });
+                var client = _httpClientFactory.CreateClient();
+                client.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", token);
+                var responseMessage = await client.DeleteAsync($"https://localhost:7140/api/Location/DeleteLocation?id={id}");
+                if (responseMessage.IsSuccessStatusCode)
+                {
+                    return RedirectToAction("Index", "AdminLocation", new { area = "Admin" });
+                }
             }
             return RedirectToAction("Index", "AdminLocation", new { area = "Admin" });
         }
@@ -79,16 +89,19 @@ namespace Ui.Areas.Admin.Controllers
         [Route("UpdateLocation/{id}")]
         public async Task<IActionResult> UpdateLocation(int id)
         {
-
-            var client = _httpClientFactory.CreateClient();
-            var responseMessage = await client.GetAsync($"https://localhost:7140/api/Location/GetLocationById?id={id}");
-            if (responseMessage.IsSuccessStatusCode)
+            var token = User.Claims.FirstOrDefault(x => x.Type == "accessToken")?.Value;
+            if (token != null)
             {
-                var jsonData = await responseMessage.Content.ReadAsStringAsync();
-                Console.WriteLine($"API Response: {jsonData}");
-                var apiResponse = JsonConvert.DeserializeObject<ApiResponse<LocationDto>>(jsonData);
-                var values = apiResponse?.Data;
-                return View(values);
+                var client = _httpClientFactory.CreateClient();
+                var responseMessage = await client.GetAsync($"https://localhost:7140/api/Location/GetLocationById?id={id}");
+                if (responseMessage.IsSuccessStatusCode)
+                {
+                    var jsonData = await responseMessage.Content.ReadAsStringAsync();
+                    Console.WriteLine($"API Response: {jsonData}");
+                    var apiResponse = JsonConvert.DeserializeObject<ApiResponse<LocationDto>>(jsonData);
+                    var values = apiResponse?.Data;
+                    return View(values);
+                }
             }
             return View();
         }
@@ -98,13 +111,18 @@ namespace Ui.Areas.Admin.Controllers
         [Route("UpdateLocation/{id}")]
         public async Task<IActionResult> UpdateLocation(LocationDto LocationDto)
         {
-            var client = _httpClientFactory.CreateClient();
-            var jsonData = JsonConvert.SerializeObject(LocationDto);
-            StringContent stringContent = new StringContent(jsonData, Encoding.UTF8, "application/json");
-            var responseMessage = await client.PostAsync("https://localhost:7140/api/Location/UpdateLocation", stringContent);
-            if (responseMessage.IsSuccessStatusCode)
+            var token = User.Claims.FirstOrDefault(x => x.Type == "accessToken")?.Value;
+            if (token != null)
             {
-                return RedirectToAction("Index", "AdminLocation", new { area = "Admin" });
+                var client = _httpClientFactory.CreateClient();
+                var jsonData = JsonConvert.SerializeObject(LocationDto);
+                StringContent stringContent = new StringContent(jsonData, Encoding.UTF8, "application/json");
+                client.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", token);
+                var responseMessage = await client.PostAsync("https://localhost:7140/api/Location/UpdateLocation", stringContent);
+                if (responseMessage.IsSuccessStatusCode)
+                {
+                    return RedirectToAction("Index", "AdminLocation", new { area = "Admin" });
+                }
             }
             return View();
         }
